cart merge

+++ app/bill/models.py
@@ -21,9 +21,8 @@ class Basket(models.Model):
         null=True,
         blank=True,
     )
-    amount = models.IntegerField(default=0)
+    amount = models.PositiveSmallIntegerField(default=0)
     order_yn = models.BooleanField(default=False)
-
 
     def __str__(self):
         return f'{self.user.username}/{self.item.item_name}/{self.amount}'
